works?

+++ src/error.rs
@@ -2,7 +2,7 @@ use std::fmt::{Display, Error as FmtError, Formatter};
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum Error {
-    OutOfBoundsUpdate { index: usize, len: usize },
+    OutOfBoundsUpdate { index: usize, len: u64 },
     OutOfBoundsIterFrom { index: usize, len: usize },
     ListFull { len: usize },
     PackedLeafFull { len: usize },
